Connected successfully

+++ cSarpProject/Models/InstitutionDetails.cs
@@ -10,7 +10,7 @@ namespace cSarpProject.Models
 {
     public class InstitutionDetails
     {
-		public Int64 Id { get; set; }
+		public Int32 Id { get; set; }
 		public string Name { get; set; }
     }
 }
